Overridable metamethods from lua (#133)
